scores: Prevent StackOverflow in frontend when duplicate score appears:
There was a cycle from score.error.score back to the score, on which
angular chokes when deep-comparing objects trees.

It was introduced in 03bbaa1152644c6da9dbb98a32ad163a11fbb928 of which
the functionality was removed in 7dd9900f868841dea6a3b3bbc036f4e6cbcb013f,
so the property can safely be removed again.